Update initiative-loi09-08.json
This is the policy set (initiative) that groups the above definitions. After you upload definitions to Azure, you will create an initiative that references them. The JSON below is a template referencing policy names after you create definitions, update policyDefinitionReferenceId with the created definition IDs, or create the initiative via CLI using the file names (as shown in the README commands